pyproject.toml
[project]
name = "ai-form"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14",
    "asyncpg>=0.30",
    "authlib>=1.5",
    "boto3>=1.35",
    "fastapi>=0.115",
    "httpx>=0.28",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1",
    "langchain-aws>=0.2",
    "langchain-core==1.2.28",
    "langgraph==1.1.6",
    "langgraph-checkpoint-postgres>=3.0.5",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0",
    "sse-starlette>=2.2",
    "uvicorn[standard]>=0.34",
]
